+++ src/trace_processor/BUILD.gn
@@ -193,7 +193,6 @@ if (enable_perfetto_trace_processor_sqlite) {
       "importers/json",
       "importers/ninja",
       "importers/perf",
-      "importers/perf_text",
       "importers/pprof",
       "importers/primes",
       "importers/proto:full",
@@ -237,6 +236,7 @@ if (enable_perfetto_trace_processor_sqlite) {
       "plugins/metadata",
       "plugins/package_lookup",
       "plugins/perf_counter",
+      "plugins/perf_text",
       "plugins/perfetto_manifest",
       "plugins/pprof_functions",
       "plugins/slice_mipmap_operator",
