+++ src/trace_processor/plugins/perf_text/perf_text_trace_parser.cc
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-#include "src/trace_processor/importers/perf_text/perf_text_trace_parser.h"
+#include "src/trace_processor/plugins/perf_text/perf_text_trace_parser.h"
 
 #include <cstdint>
 
 #include "src/trace_processor/importers/common/process_tracker.h"
 #include "src/trace_processor/importers/common/stack_profile_tracker.h"
-#include "src/trace_processor/importers/perf_text/perf_text_event.h"
+#include "src/trace_processor/plugins/perf_text/perf_text_event.h"
 #include "src/trace_processor/storage/trace_storage.h"
 #include "src/trace_processor/tables/profiler_tables_py.h"
 #include "src/trace_processor/types/trace_processor_context.h"
