+++ src/trace_processor/trace_processor_impl.cc
@@ -119,6 +119,7 @@
 #include "src/trace_processor/plugins/metadata/metadata.h"
 #include "src/trace_processor/plugins/package_lookup/package_lookup.h"
 #include "src/trace_processor/plugins/perf_counter/perf_counter.h"
+#include "src/trace_processor/plugins/perf_text/perf_text.h"
 #include "src/trace_processor/plugins/perfetto_manifest/perfetto_manifest.h"
 #include "src/trace_processor/plugins/pprof_functions/pprof_functions.h"
 #include "src/trace_processor/plugins/slice_mipmap_operator/slice_mipmap_operator.h"
@@ -353,6 +354,7 @@ TraceProcessorImpl::TraceProcessorImpl(const Config& cfg)
   metadata::RegisterPlugin();
   package_lookup::RegisterPlugin();
   perf_counter::RegisterPlugin();
+  perf_text_importer::RegisterPlugin();
   perfetto_manifest::RegisterPlugin();
   pprof_functions::RegisterPlugin();
   slice_mipmap_operator::RegisterPlugin();
@@ -453,7 +455,6 @@ TraceProcessorImpl::TraceProcessorImpl(const Config& cfg)
   reg.Register(CreateSimpleperfProtoImporter());
   reg.Register(CreateTarImporter());
   reg.Register(CreatePrimesImporter());
-  reg.Register(CreatePerfTextImporter());
 
   // Force initialization of heap graph tracker.
   //
